Move UserUtils to firebase folder and center mobile title text
