+++ src/css/Header.scss
@@ -51,7 +51,7 @@
   #header-title {
     text-align: center;
     font-weight: 800;
-    padding: 6px 0;
+    padding: 6px 0 6px 50px;
   }
 }
